# optimized(monitor): query metrics time